Add time_incremental parameter config.
